requirements.txt
# pip install -r requirements.txt

Django==1.9.3
argparse==1.3.0
django-bootstrap3==7.0.0
django-cron==0.4.6
pbr==0.10.8
rpi.gpio==0.6.2
stevedore==1.3.0
virtualenv==12.0.7
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.1
wsgiref==0.1.2
